Rename all jsx files
