+++ frontend/src/components/TableUser/TableUsers.jsx
@@ -3,7 +3,7 @@ import "./tabela-korisnici.css";
 import { Link } from "react-router-dom";
 import { ModalWindow } from "../ModalWindow/ModalWindow";
 
-export const TabelaKorisnici = ({ isDriver, isEdit }) => {
+export const TableUsers = ({ isDriver, isEdit }) => {
     const [korisnici, setKorisnici] = useState([]);
     const [isModalOpen, setIsModalOpen] = useState(false);
     const [modalData, setModalData] = useState({});
